+++ PrivateFrameworks/Symbolication.framework/VMUCallTreeNode.h
@@ -14,7 +14,7 @@
         NSString *thePseudoName; 
         VMUCallTreeNode *theChild; 
         VMUCallTreeNode **theChildren; 
-    unsigned int address;
+    unsigned long long address;
     unsigned int count;
     unsigned int flags;
     NSString *name;
@@ -35,7 +35,7 @@
 - (void)addStackEntry:(id)arg1 symbolicator:(struct _CSTypeRef { unsigned int x1; unsigned int x2; })arg2 sampler:(id)arg3 numBytes:(unsigned int)arg4 options:(unsigned int)arg5 uniqueStrings:(id)arg6 addressToSymbolNameMap:(id)arg7 threadPortToNameMap:(id)arg8 dispatchQueueSerialNumToNameMap:(id)arg9;
 - (char *)addSubTreeFromFile:(struct __sFILE { char *x1; int x2; int x3; short x4; short x5; struct __sbuf { char *x_6_1_1; int x_6_1_2; } x6; int x7; void *x8; int (*x9)(); int (*x10)(); int (*x11)(); int (*x12)(); struct __sbuf { char *x_13_1_1; int x_13_1_2; } x13; struct __sFILEX {} *x14; int x15; unsigned char x16[3]; unsigned char x17[1]; struct __sbuf { char *x_18_1_1; int x_18_1_2; } x18; int x19; long long x20; }*)arg1 withIndent:(int)arg2 withLine:(char *)arg3 withLen:(unsigned int*)arg4;
 - (void)addTraceEvent:(id)arg1 forTraceData:(id)arg2;
-- (unsigned int)address;
+- (unsigned long long)address;
 - (id)browserName;
 - (int)compare:(id)arg1;
 - (int)compareNames:(id)arg1;
@@ -47,8 +47,8 @@
 - (id)filterOutSymbols:(id)arg1;
 - (id)filterOutWaiting;
 - (id)findNodeMatching:(id)arg1 searchForward:(BOOL)arg2 ignoreCase:(BOOL)arg3 wholeWords:(BOOL)arg4;
-- (id)findOrAddChildWithName:(id)arg1 address:(unsigned int)arg2 compareSymbolNames:(BOOL)arg3;
-- (id)findOrAddChildWithName:(id)arg1 address:(unsigned int)arg2;
+- (id)findOrAddChildWithName:(id)arg1 address:(unsigned long long)arg2 compareSymbolNames:(BOOL)arg3;
+- (id)findOrAddChildWithName:(id)arg1 address:(unsigned long long)arg2;
 - (id)fullOutputWithThreshold:(unsigned int)arg1;
 - (id)invertedNode;
 - (BOOL)isPseudo;
